Avoid multiple concurrent observe() calls with Transformations.switchMap
Instead of starting a new `observe` call every time the query button is pressed, we should just update our current source of data by using `Transformations.switchMap`. We store the user's query in a `SavedStateHandle` to ensure that we handle process death and recreation properly (and to take advantage of the `getLiveData()` method to drive our `switchMap`.

Test: tests still pass

Fixes #759